// apps/frontend/src/app/api/generated/models/scope-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ScopeResponse {
  active?: boolean;
  executingUnit?: string;
  executingUnitId?: number;
  id?: number;
  institution?: string;
  institutionId?: number;
  role?: 'ADMINISTRADOR_PIIP' | 'CONSULTA_EXTERNA';
  validFrom?: string;
  validUntil?: string;
  version?: number;
}